changing-base-project
we changed the base project to react-hooks-global-states, react-hooks-global-states is equal to react-native-global-state-hooks version 5.0.15

we separate it into a different project for allowing this react-native version to be focus on native.

by doing this above now we implemented persist storage in build avoiding having to create custom implementations on the consumers.